Add vd_toml tests
